# person-service/openapi/person-api.yaml
openapi: 3.0.0
info:
  title: Person API
  description: REST API для взаимодействия с Person сервисом
  version: 1.0.0
servers:
  - url: http://localhost:8082

paths:
  /v1/persons:
    post:
      summary: Регистрация пользователя
      operationId: "registration"
      tags:
        - person-service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndividualWriteDto'
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualWriteResponseDto'

    get:
      summary: Получение списка пользователей
      operationId: "findAllByEmail"
      tags:
        - person-service
      parameters:
        - in: query
          name: email
          required: false
          schema:
            type: array
            items:
              type: string
              format: email
          style: form
          explode: true
      responses:
        '200':
          description: Успешное получение списка пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualPageDto'
  
  /v1/persons/{id}:
    
    get:
      summary: Получение пользователя по Id
      operationId: "findById"
      tags:
        - person-service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Успешное получение пользователя по Id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualDto'
    
    put:
      summary: Изменение пользователя
      operationId: "update"
      tags:
        - person-service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndividualWriteDto'
      responses:
        '201':
          description: Успешно обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndividualWriteResponseDto'
  
    delete:
      summary: Удаление пользователя (soft)
      operationId: "delete"
      tags:
        - person-service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешно удалено
    
  /v1/persons/compensate-registration/{id}:
    delete:
      summary: Удаление пользователя (hard) (Компенсация регистрации пользователя SAGA)
      operationId: "compensateRegistration"
      tags:
        - person-service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешно удалено
  

components:
  schemas:
    IndividualWriteResponseDto:
      type: object
      required: [ id ]
      properties:
        id:
          type: string
      
    IndividualWriteDto:
      type: object
      required: [
        email,
        firstName,
        lastName,
        passport_number,
        phone_number,
        address,
      ]
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        passport_number:
          type: string
        phone_number:
          type: string
        address:
          $ref: '#/components/schemas/AddressWriteDto'
        
    AddressWriteDto:
      type: object
      required: [ ]
      properties:
        address:
          type: string
        zip_code:
          type: string
        city:
          type: string
        country_code:
          type: string
    
    IndividualDto:
      type: object
      required: [
        email,
        firstName,
        lastName,
        passport_number,
        phone_number,
        address,
        active
      ]
      properties:
        firstName:
          type: string
          format: email
        lastName:
          type: string
          format: email
        email:
          type: string
          format: email
        passport_number:
          type: string
        phone_number:
          type: string
        address:
          $ref: '#/components/schemas/AddressDto'
        active:
          type: boolean

    AddressDto:
      type: object
      required: [ ]
      properties:
        address:
          type: string
        zip_code:
          type: string
        city:
          type: string
        country_code:
          type: string

    IndividualPageDto:
      type: object
      required: [ ]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/IndividualDto'
  